GraphGenerator working
